requirements.txt
apache-airflow-providers-google
pyarrow
python-dotenv
dbt-core
dbt-bigquery